Move CounterApp middleware to global middleware to how it composes with the transactor
